Removed duplicated code. We have created a 'parent' class called 'Country' and the file 'CountryFunctions' with the common functions
