--- wiki-countries/src/components/CountryDetailItem.js
@@ -1,17 +1,0 @@
-import React, { Component } from "react";
-import { Link } from 'react-router-dom';
-
-export class CountryDetailItem extends Component {
-
-
-	handleClick = (e) => {
-		this.props.onBorder(this.props.id)
-	}
-
-	render() {
-		return (
-			
-		)
-	}
-
-}
